build() function instantiates template
With the build method an actual template can be instatiated. A dist directory is created (by default 'dist/') and the produced content comes from the instantiation of the template.

Currently, no files from the user config is kept. In the future, user defined images will be kept and copied with the build, so that a user will be able to insert their own images.