Added get_defaults function
Used to retrieve default dictionaries. This will soon override the
current defaults management (where the dictionaries are declared as
constants in the code).

The new behaviour loads defaults from a global file
(templates/defaults.yaml), and updates them with template specific
defaults, which are defined by the template's creator in _defaults.yaml.